feat: [#286] 유저페이지 헤더 모달 라우팅
라우트 모달 내부 구조 변경
