+++ front/src/pages/MainPage.tsx
@@ -1,107 +1,41 @@
-import React, { useRef, useState } from 'react';
-import Header from '@components/Header/Header';
-import FeedContainer from '@components/Feed/FeedContainer';
-import FeedFAB from '@components/Feed/FeedFAB';
-import Explore from '@components/Explore/Explore';
+import React from 'react';
 import styled from 'styled-components';
-import HabitatPreview from '@components/Habitat/HabitatPreview';
-import useSideNavi from '@src/hooks/useSideNavi';
-import { Palette } from '@lib/styles/Palette';
-import { flexBox } from '@lib/styles/mixin';
-import useHabitatInfo from '@hooks/useHabitatInfo';
-import MagicNumber from '@src/lib/styles/magic';
+import { Route } from 'react-router-dom';
+import UserArticle from '@src/pages/UserArticle';
+import MainArticle from '@src/pages/MainArticle';
+import Header from '@components/Header/Header';
 import { useUserState } from '@src/contexts/UserContext';
-import useValidateUser from '@src/hooks/useValidateUser';
-import { ScrollProvider } from '@src/contexts/ScrollContext';
+import useValidateUser from '@hooks/useValidateUser';
+import useSideNavi from '@hooks/useSideNavi';
+import useHabitatInfo from '@hooks/useHabitatInfo';
 
-const MainPage = () => {
-  const userState = useUserState();
-  useValidateUser(userState);
-
-  const [mode, setMode] = useState<'feed' | 'explore'>('feed');
-  const feedModeRef = useRef<HTMLDivElement>(null);
-  const { getCurHabitat, handleNextHabitat, handlePrevHabitat, error } = useSideNavi(userState.data?.habitatId ?? 2);
-
-  const { habitatInfo } = useHabitatInfo(getCurHabitat());
-
-  const toggleMode = () => {
-    if (feedModeRef.current) {
-      feedModeRef.current.style.opacity = '0';
-      feedModeRef.current.style.zIndex = '1';
-      setTimeout(() => {
-        setMode(mode === 'feed' ? 'explore' : 'feed');
-        if (feedModeRef.current) {
-          feedModeRef.current.style.opacity = '1';
-        }
-      }, 700);
-    }
-  };
-
-  const getFeedFloatingPos = () => (window.innerWidth + parseInt(MagicNumber.FEED_SECTION_WIDTH)) / 2 + 10;
-  const getExploreFloatingPos = () => 0;
-
-  if (error) {
-    return <div>잘못된 경로입니다!</div>;
-  }
-
-  return (
-    <MainPageBlock>
-      <Header habitatInfo={habitatInfo} />{' '}
-      <ScrollProvider>
-        <MainContentsDiv ref={feedModeRef}>
-          {
-            {
-              feed: (
-                <>
-                  <FeedContainer habitatInfo={habitatInfo} curHabitatId={getCurHabitat()} />
-                  <FeedFAB mode={mode} getPosFunc={getFeedFloatingPos} toggleMode={toggleMode} />
-                  <HabitatPreview habitat={getCurHabitat(1)} onClick={handleNextHabitat} side={'right'} />
-                  <HabitatPreview habitat={getCurHabitat(-1)} onClick={handlePrevHabitat} side={'left'} />
-                </>
-              ),
-              explore: (
-                <>
-                  <Explore habitatInfo={habitatInfo} />
-                  <FeedFAB mode={mode} getPosFunc={getExploreFloatingPos} toggleMode={toggleMode} />
-                </>
-              ),
-            }[mode]
-          }
-        </MainContentsDiv>
-      </ScrollProvider>
-      <EmptyStyleDiv color={habitatInfo?.habitat.color} />
-    </MainPageBlock>
-  );
-};
-
-export default MainPage;
-
-const MainPageBlock = styled.div`
-  ${flexBox(null, null, 'column')};
+const MainPageDiv = styled.div`
   overflow: hidden;
   position: relative;
   width: 100vw;
   height: 100vh;
-  background-color: ${Palette.BACKGROUND_GRAY};
 `;
 
-const MainContentsDiv = styled.div`
-  opacity: 1;
-  transition: opacity 0.5s linear;
-  position: relative;
-  display: inherit;
-  width: 100%;
-  height: calc(100% - ${MagicNumber.HEADER_HEIGHT});
-`;
+const MainPage = () => {
+  const userState = useUserState();
+  useValidateUser(userState);
+  const { getCurHabitat, handleNextHabitat, handlePrevHabitat, error } = useSideNavi(userState.data?.habitatId ?? 2);
 
-const EmptyStyleDiv = styled.div<{ color: string | undefined }>`
-  width: 500px;
-  margin: auto;
-  left: 0;
-  right: 0;
-  background-color: ${(props) => (props.color !== undefined ? props.color : Palette.PINK)};
-  height: calc(100% - ${MagicNumber.HEADER_HEIGHT});
-  top: ${MagicNumber.HEADER_HEIGHT};
-  position: absolute;
-  overflow-y: scroll;
-`;
+  const { habitatInfo } = useHabitatInfo(getCurHabitat());
+
+  return (
+    <MainPageDiv>
+      <Header habitatInfo={habitatInfo} />
+
+      <Route exact path="/user/:id/(modal/.*)?">
+        <UserArticle />
+      </Route>
+
+      <Route exact path="/(modal/.*)?">
+        <MainArticle getCurHabitat={getCurHabitat} handleNextHabitat={handleNextHabitat} handlePrevHabitat={handlePrevHabitat} error={error} habitatInfo={habitatInfo} />
+      </Route>
+    </MainPageDiv>
+  );
+};
+
+export default MainPage;
